test(desktop): Cover System Proxy config patches
